Tentative fix for LOF experiments: fix bug that MIX after clear_row causes exception.
Calling recommender::euclid_lsh::clear_row between get_diff and set_mixed_and_clear_diff causes inconsistencies of lsh_master_table(_diff)_ and lsh_table_,
which ends up with failure of MIX of anomaly::lof which tries to look up the (kdist, lrd) value of some already-cleared neighbor of a target row.
This change is a work-around for this problem.
The main idea is that, under current algorithms, diff should be cleared at get_diff rather than set_mixed to maintain consistencies of indexes and models.
This requires get_diff to be non-const, which is probably prohibitive to the current framework, so this change is tentative for experiments and NOT FOR MERGE TO MASTER.