Use unweighted mean of neighbors for complete_row to fix #413
This commit changes the behavior of complete_row.
It used a mean of neighbors weighted by the similarities with the query.
However, euclid_lsh returns a negated distance as a similarity.
It causes buggy results of complete_row.
So in #413, we decided to stop using weighting on complete_row, since in some cases as euclid_lsh, we cannot define an appropriate similarity for weighting.
This change is the fix for it.