Save recommender's original matrix by WORKAROUND (fix #537).
This change includes a workaround that uses a dummy mixable to correctly save/load the original data matrix of recommender_base.
We currenctly do not have distinction between mixables and data structures to be saved/loaded.
Before this fix, only mixables registered to mixable_holder are saved/loaded.
It causes lost of the original data matrix on saving recommender_base.

This change tentatively resolves this degradation by adding a dummy mixable that just saves the matrix and does nothing on MIX.
It should be fixed by more elegant way in the future.